Backend: Fixed logical errors in map editor

+++ server/services/shapeService.js
@@ -1,3 +1,4 @@
+const Slot = require("../models/slotModel");
 const Shape = require("../models/shapeModel");
 const mongoose = require("mongoose");
 
@@ -13,8 +14,18 @@ exports.modifyMapLayout = async (data) => {
   const frontendIds = shapes.filter((s) => s._id).map((s) => s._id);
 
   const shapesToDelete = existingIds.filter((id) => !frontendIds.includes(id));
+  const slotsToDelete = await Slot.find({
+    slotId: { $in: shapesToDelete },
+  });
+  const assignedSlots = slotsToDelete.filter((s) => s.assignedStudentId);
+
+  if (assignedSlots.length > 0) {
+    throw new Error("You cannot delete assigned slot");
+  }
+
   if (shapesToDelete.length > 0) {
     await Shape.deleteMany({ _id: { $in: shapesToDelete } });
+    await Slot.deleteMany({ slotId: { $in: shapesToDelete } });
   }
 
   const updatedShapes = [];
@@ -27,6 +38,14 @@ exports.modifyMapLayout = async (data) => {
       result = await Shape.findByIdAndUpdate(_id, rest, { new: true });
     } else {
       result = await Shape.create(rest);
+
+      if (rest.metadata?.type === "slot") {
+        await Slot.create({
+          assignedStudentId: null,
+          slotId: result._id,
+          status: "available",
+        });
+      }
     }
 
     if (result) updatedShapes.push(result);
